Rename and move network files and components
